Git catchup: changes in runtime model, tree logic
Playing a little bit of git catchup. Instead of the old, slow,
inevitably I/O bound 'application software' approach, pivoted to making
DMCTS handlers easily used in application-specific lambda deployments.
It seems like it'll end up being far more clean, and involve far less
yak shaving than defining an interface for external application specific
code.

First pass at the SearchTree code, with functions to sample random
leafs, as well as random spines. Instead of a simple averaging function,
made an "aggregate" function a part of the WeightAble class, so users
can define their own aggregation semantics (Max weight? Min weight?
Numeric average?). Works out well, instead of trying to fight the type
system into letting me use a static averaging function on the
lambda-side.

Defined DMCTSRequest and DMCTSResponse structures to wrap our labels in.
Carry metadata about the MCTS. Currently, lets users define their
desired sampling method (leaf or spine), the number of samples, the root
node, and the seed for the Monte Carlo method. I like the inclusion of a
deterministic seed, as it'll allow us to both seed randomly when we're
using the system in practice, and pin the seed to certain values for
unit testing.

Implemented the beginnings of a knapsack problem proof of concept. Since
Santorini is complex, I thought it better to start off with a simpler
problem. DMCTS doesn't _really_ fit Knapsack problems if we're looking
for a global maximum, but it's challenging enough that we can use it to
push the system, and simple enough to be a good proof-of-concept. It
should prove as a good benchmark.